lindex/requirements.txt
llama-index-embeddings-azure-openai
llama-index-llms-azure-openai
llama-index
llama-index-core
python-dotenv 
openai